requirements.txt
# AnimalShortsBot — All Free Dependencies
# Install: pip install -r requirements.txt

# AI Script Generation
anthropic>=0.25.0
httpx>=0.27.0

# Text-to-Speech (free, no API key)
edge-tts>=6.1.9          # Microsoft Edge TTS — best quality, completely free
gTTS>=2.5.0              # Google TTS — free fallback

# Footage Fetching
httpx>=0.27.0            # async HTTP (already listed above)

# YouTube Upload
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Scheduling & Timezone
apscheduler>=3.10.4
pytz>=2024.1

# Utilities
asyncio-throttle>=1.0.2
python-dotenv>=1.0.1

# Note: FFmpeg is NOT a pip package — install separately:
# Railway: add "nixpacks.toml" (see README)
# Render: uses Docker — see Dockerfile
